reduce streamProcessor references in BufferController, add ADD_INBAND_EVENTS_REQUESTED event
